feat(api): first draft of timeout handling
