+++ api/Cargo.toml
@@ -20,7 +20,7 @@ network_bridge = "0.1.1"
 network-interface = "1.0.0"
 cidr = "0.2.1"
 rand = "0.8.4"
-tokio = { version = "1.0", features = ["rt-multi-thread", "macros"] }
+tokio = { version = "1.0", features = ["rt-multi-thread", "macros", "process"] }
 tonic = { version = "0.9.2", features = ["transport"] }
 prost = "0.11.9"
