+++ api/src/vm_manager/vmm/mod.rs
@@ -1,11 +1,18 @@
 pub mod grpc_definitions;
 pub mod grpc_server;
+mod net;
 
-use std::{error::Error as STDError, fmt::Display};
+use std::{error::Error as STDError, fmt::Display, str::FromStr};
 
-use cidr::IpInet;
+use cidr::{IpInet, Ipv4Inet};
+use log::{debug, error, info, trace};
 use lumper::VMM;
 use tokio::task::JoinHandle;
+use uuid::Uuid;
+
+use crate::{model::LanguageSettings, vm_manager::state::VMState};
+
+use super::state::LambdoState;
 
 #[derive(Debug)]
 pub enum Error {
@@ -18,6 +25,8 @@ pub enum Error {
     VmNotFound,
     VmAlreadyEnded,
     GrpcError,
+    ExecutionError,
+    Timeout,
 }
 
 impl STDError for Error {}
@@ -34,6 +43,8 @@ impl Display for Error {
             Error::VmNotFound => write!(f, "VM not found"),
             Error::VmAlreadyEnded => write!(f, "VM already ended"),
             Error::GrpcError => write!(f, "GRPC error"),
+            Error::ExecutionError => write!(f, "Execution error"),
+            Error::Timeout => write!(f, "Timeout"),
         }
     }
 }
@@ -80,4 +91,58 @@ pub fn run(opts: VMMOpts) -> Result<JoinHandle<Result<(), Error>>, Error> {
     Ok(tokio::task::spawn_blocking(move || {
         vmm.run(true).map_err(Error::VmmRun)
     }))
+}
+
+pub async fn run_vm(
+    state: &mut LambdoState,
+    language_settings: &LanguageSettings,
+    reserved: bool,
+) -> Result<String, Error> {
+    let ip = net::find_available_ip(state).await.map_err(|e| {
+        error!("Error while finding available IP address: {:?}", e);
+        Error::NoIPAvalaible
+    })?;
+    let uuid = Uuid::new_v4().to_string();
+
+    let config = &state.config;
+    // Safe since we checked the validity of the address before
+    let host_ip = Ipv4Inet::from_str(&config.api.bridge_address).unwrap();
+    let tap_name = format!("tap-{}", uuid[0..8].to_string());
+
+    let opts: VMMOpts = VMMOpts {
+        kernel: config.vmm.kernel.clone(),
+        cpus: 1,
+        memory: 1024,
+        console: None,
+        socket: None,
+        initramfs: Some(language_settings.initramfs.clone()),
+        tap: Some(tap_name.clone()),
+        ip: Some(IpInet::V4(ip)),
+        gateway: Some(host_ip.address().to_string()),
+    };
+
+    trace!("Creating VMState");
+    let mut vm_state = VMState::new(
+        uuid.clone(),
+        opts.clone(),
+        language_settings.clone(),
+        state.channel.0.clone(),
+        reserved,
+    );
+
+    info!(
+        "Starting execution for {:?}, (language: {}, version: {})",
+        &uuid, language_settings.name, language_settings.version
+    );
+    debug!("Launching VMM with options: {:?}", opts);
+    vm_state.vm_task = Some(run(opts)?);
+
+    debug!("Adding interface to bridge");
+    net::add_interface_to_bridge(&tap_name, &*state).map_err(|e| {
+        error!("Error while adding interface to bridge: {:?}", e);
+        Error::NoIPAvalaible
+    })?;
+    state.vms.push(vm_state);
+
+    Ok(uuid.clone())
 }
